Add flag to limit 'unlimited' searches.
This is to prevent the system from freezing up as it searches
for and displays hundreds of thousands of songs that contain
a simple search string. In fact, empty string matches every song.

This constant is hard-coded at compile-time, and a more sophisticated
solution would allow it to be set in rdadmin.