Operating Mode UI and state changes.
Three interlocking changes to improve the usability of rdairplay
by DJs. These changes stem from the operating mode doing two things:
changing the playback mode and activating automated playback. With
these changes, the button only controls playback mode, and WMFO uses
buttons on the SoundPanel to control automation. We co-opt the on-air
flag to be the automation-is-off-flag.

To do this, we:
* Change the title of 'Operating Mode' to 'Playback Mode'.
    'Automatic' becomes 'Continuous'. 'Manual' becomes 'Single'.
* Modify the state machine to make LiveAssist inaccessible. We
    ignore the logStartMode and always start in Continuous.
* Do not transition to Single (Manual) mode while automation is on.
    We do this by keeping track of the on-air flag's state (it doesn't
    seem to be directly accessible). At WMFO, turning on automation
    always sets Continuous mode, but without this change the user
    could click back to Single mode, leaving the station broadcasting
    silence unattended after the current song finishes.